chore: level1 refactor done :hammer:
